+++ src/main/java/roomescape/exception/dto/ErrorResponse.java
@@ -1,4 +1,4 @@
-package roomescape.exception;
+package roomescape.exception.dto;
 
 public record ErrorResponse(
         String message
